pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "certimate_webhook"
version = "1.0.0"
description = "SSL certificate webhook for automated deployment"
authors = [{name = "Admin", email = "admin@example.com"}]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "loguru>=0.7.2",
    "ansible-runner>=2.3.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.6",
]

[tool.ruff]
target-version = "py313"
line-length = 88
select = ["E", "W", "F", "I", "B", "C4", "ARG", "SIM"]
ignore = ["E501", "W291", "W292", "W293"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = [
    "*.py",
    "config.yml"
]

[project.scripts]
certimate_webhook = "main:main"